HSEARCH-2772 Move the JsonHelper test util to hibernate-search-testing
This way, we don't have to duplicate it in both the elsaticsearch module
and the integrationtest/elasticsearch module.
