+++ integrationtest/elasticsearch/src/test/java/org/hibernate/search/elasticsearch/test/Elasticsearch5IndexMappingIT.java
@@ -19,12 +19,12 @@ import org.hibernate.search.FullTextSession;
 import org.hibernate.search.Search;
 import org.hibernate.search.elasticsearch.ElasticsearchProjectionConstants;
 import org.hibernate.search.elasticsearch.ElasticsearchQueries;
-import org.hibernate.search.elasticsearch.testutil.JsonHelper;
 import org.hibernate.search.elasticsearch.testutil.TestElasticsearchClient;
 import org.hibernate.search.elasticsearch.testutil.junit.SkipBelowElasticsearch50;
 import org.hibernate.search.elasticsearch.util.impl.ElasticsearchDateHelper;
 import org.hibernate.search.query.engine.spi.QueryDescriptor;
 import org.hibernate.search.test.SearchTestBase;
+import org.hibernate.search.test.util.JsonHelper;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
