+++ integrationtest/elasticsearch/src/test/java/org/hibernate/search/elasticsearch/test/ElasticsearchSchemaCreateStrategyIT.java
@@ -6,8 +6,8 @@
  */
 package org.hibernate.search.elasticsearch.test;
 
-import static org.hibernate.search.elasticsearch.testutil.JsonHelper.assertJsonEquals;
-import static org.hibernate.search.elasticsearch.testutil.JsonHelper.assertJsonEqualsIgnoringUnknownFields;
+import static org.hibernate.search.test.util.JsonHelper.assertJsonEquals;
+import static org.hibernate.search.test.util.JsonHelper.assertJsonEqualsIgnoringUnknownFields;
 
 import java.util.HashMap;
 import java.util.Map;
