+++ integrationtest/elasticsearch/src/test/java/org/hibernate/search/elasticsearch/test/ToElasticsearchIT.java
@@ -7,7 +7,7 @@
 package org.hibernate.search.elasticsearch.test;
 
 import static org.fest.assertions.Assertions.assertThat;
-import static org.hibernate.search.elasticsearch.testutil.JsonHelper.assertJsonEquals;
+import static org.hibernate.search.test.util.JsonHelper.assertJsonEquals;
 
 import java.util.Calendar;
 import java.util.Date;
