+++ testing/src/main/java/org/hibernate/search/test/util/JsonHelper.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.testutil;
+package org.hibernate.search.test.util;
 
 import org.hibernate.search.exception.AssertionFailure;
 import org.json.JSONException;
